+++ src/api/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Route handler to use http listener
+- Command handler
+
 ## [v0.0.1] - 2023-11-13
 
 ### Added